feat: add nosessionactive button
